+++ src/bos/operators/utils/clients/bos/sessions_status.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2021-2024 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021-2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -21,15 +21,17 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+from functools import partial
 import json
 import logging
 
 from bos.common.tenant_utils import get_new_tenant_header
 from bos.common.utils import requests_retry_session
 from .base import BASE_ENDPOINT, log_call_errors
+from .base import check_bos_response as _check_bos_response
 
 LOGGER = logging.getLogger('bos.operators.utils.clients.bos.sessions_status')
-
+check_bos_response = partial(_check_bos_response, logger=LOGGER)
 
 class SessionStatusEndpoint:
     ENDPOINT = 'sessions'
@@ -44,7 +46,7 @@ class SessionStatusEndpoint:
         session = requests_retry_session()
         LOGGER.debug("GET %s for tenant=%s", url, tenant)
         response = session.get(url, headers=get_new_tenant_header(tenant))
-        response.raise_for_status()
+        check_bos_response(response)
         item = json.loads(response.text)
         return item
 
@@ -58,6 +60,6 @@ class SessionStatusEndpoint:
         url = self.base_url + '/' + session_id + '/status'
         LOGGER.debug("POST %s for tenant=%s", url, tenant)
         response = session.post(url, headers=get_new_tenant_header(tenant))
-        response.raise_for_status()
+        check_bos_response(response)
         items = json.loads(response.text)
         return items